requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
requirements-parser==0.11.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
httpx==0.27.0
aiocache==0.12.3
pytest==8.3.5
pytest-asyncio==0.26.0